% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACATO.R
\name{ACATO}
\alias{ACATO}
\title{ACAT function by Yaowu Liu, with changes}
\usage{
ACATO(p)
}
\arguments{
\item{p:}{integrated in function of mBAT_SVD.R, using P-value of mBAT and fastBAT}
}
\value{
P-value of mBAT-combo
}
\description{
combine the test results of mBAT and fastBAT by the aggregated Cauchy combination method (ACAT)
}
