% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_flip_gwas_beta.R
\name{allele_flip_gwas_beta}
\alias{allele_flip_gwas_beta}
\title{functions for allele flip and dif_AF1 >= 0.2}
\usage{
allele_flip_gwas_beta(gwas, info_LD)
}
\arguments{
\item{info_LD}{}
}
\description{
functions for allele flip and dif_AF1 >= 0.2
}
