% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBAT_combo.R
\name{mBAT_combo}
\alias{mBAT_combo}
\title{mBAT-combo, combining P-values of mBAT and fastBAT to gain the highest power.}
\usage{
mBAT_combo(
  bim_file,
  map_file,
  assoc_file,
  LD_path,
  result_path,
  fastBAT_output,
  prop,
  gene_annotate
)
}
\arguments{
\item{fastBAT_output}{}
}
\description{
mBAT-combo, combining P-values of mBAT and fastBAT to gain the highest power.
}
