% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbat_SVD.R
\name{mbat_SVD}
\alias{mbat_SVD}
\title{Function for mbat and mBAT-combo calculation with SVD}
\usage{
mbat_SVD(gwas_noqc_final, eig, prop, fastBAT_call, z, chr_name, ID)
}
\arguments{
\item{ID}{}
}
\description{
Function for mbat and mBAT-combo calculation with SVD
}
