% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_gwas_op.R
\name{qc_gwas_op}
\alias{qc_gwas_op}
\title{Function for generating output of gwas after QC}
\usage{
qc_gwas_op(gwas, qc_snp_list)
}
\arguments{
\item{qc_snp_list}{}
}
\description{
Function for generating output of gwas after QC
}
