% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_LD_op.R
\name{qc_LD_op}
\alias{qc_LD_op}
\title{Function for generating output of LD after QC}
\usage{
qc_LD_op(gwas, LD, qc_snp_list)
}
\arguments{
\item{qc_snp_list}{}
}
\description{
Function for generating output of LD after QC
}
