% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLDMbin.R
\name{ReadLDMbin}
\alias{ReadLDMbin}
\title{Function for reading in LD file generated by GCTB}
\usage{
ReadLDMbin(prefix)
}
\arguments{
\item{prefix}{}
}
\description{
Function for reading in LD file generated by GCTB
}
