% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp2gene.R
\name{snp2gene}
\alias{snp2gene}
\title{Map SNPs to genes function by Jaehyun Joo, with changes
Annotate SNPs onto their arbitrary genomic regions to perform set-based association tests.}
\usage{
snp2gene(
  info_snp,
  info_gene,
  window_start = 50L,
  window_end = 50L,
  only_sets = FALSE
)
}
\arguments{
\item{only_sets}{}
}
\description{
Map SNPs to genes function by Jaehyun Joo, with changes
Annotate SNPs onto their arbitrary genomic regions to perform set-based association tests.
}
